[Refactor] 전체 리뷰 수 및 오류, 스타일 수정
# 287
